requirements.txt
pyttsx3~=2.90
pywhatkit~=5.3
wikipedia~=1.4.0
pyjokes~=0.6.0
streamlit~=1.4.0
Pillow~=7.2.0